+++ app/main/controller/langid_controller.py
@@ -4,7 +4,7 @@ import hashlib
 import json
 import importlib
 import tenacity
-from app.main.lib import redis
+from app.main.lib import redis_client
 
 from twitter_text import extract_urls_with_indices, extract_emojis_with_indices
 
@@ -25,7 +25,7 @@ class LangidResource(Resource):
         if 'provider' in request.json: provider = request.json['provider']
 
         # Read from cache first.
-        r = redis.get_client()
+        r = redis_client.get_client()
         key = 'langid:' + provider + ':' + hashlib.md5(text.encode('utf-8')).hexdigest()
         try:
             result = json.loads(r.get(key))
