+++ app/main/lib/media_crud.py
@@ -13,7 +13,7 @@ from app.main import db
 from app.main.model.video import Video
 from app.main.lib.presto import Presto, PRESTO_MODEL_MAP
 from app.main.lib.similarity_helpers import drop_context_from_record
-from app.main.lib import redis
+from app.main.lib import redis_client
 
 def merge_dict_lists(list1, list2):
     """
@@ -167,7 +167,6 @@ def get_blocked_presto_response(task, model, modality):
         return obj, temporary, get_context_for_search(task), obj.canned_response
 
 def get_async_presto_response(task, model, modality):
-    app.logger.error(f"Task at this point is {task}")
     obj, temporary = get_object(task, model)
     context = [get_context_for_search(task)]
     callback_url =  Presto.add_item_callback_url(app.config['ALEGRE_HOST'], modality)
