+++ app/test/test_sync_similarity.py
@@ -11,7 +11,7 @@ from app.main.lib.shared_models.shared_model import SharedModel
 from unittest.mock import patch
 from app.main.model.audio import Audio
 from app.main.lib.shared_models.audio_model import AudioModel
-from app.main.lib import redis
+from app.main.lib import redis_client
 class TestSyncSimilarityBlueprint(BaseTestCase):
     def setUp(self):
         super().setUp()
@@ -25,7 +25,7 @@ class TestSyncSimilarityBlueprint(BaseTestCase):
     def test_audio_basic_http_responses_with_doc_id(self):
         url = 'file:///app/app/test/data/test_audio_1.mp3'
         with patch('requests.post') as mock_post_request:
-            r = redis.get_client()
+            r = redis_client.get_client()
             r.delete(f"audio_1c63abe0-aeb4-4bac-8925-948b69c32d0d")
             r.lpush(f"audio_1c63abe0-aeb4-4bac-8925-948b69c32d0d", json.dumps({"body": {"result": {"hash_value": [1,2,3]}}}))
             mock_response = Mock()
@@ -60,7 +60,7 @@ class TestSyncSimilarityBlueprint(BaseTestCase):
     def test_audio_basic_http_responses(self):
         url = 'http://example.com/blah.mp3'
         with patch('requests.post') as mock_post_request:
-            r = redis.get_client()
+            r = redis_client.get_client()
             r.delete(f"audio_1c63abe0-aeb4-4bac-8925-948b69c32d0d")
             r.lpush(f"audio_1c63abe0-aeb4-4bac-8925-948b69c32d0d", json.dumps({"body": {"result": {"hash_value": [1,2,3]}}}))
             mock_response = Mock()
@@ -96,7 +96,7 @@ class TestSyncSimilarityBlueprint(BaseTestCase):
         with patch('requests.post') as mock_post_request:
             with patch('app.main.lib.image_similarity.execute_command') as mock_db_response:
                 mock_db_response.return_value = [(1, "1c63abe0-aeb4-4bac-8925-948b69c32d0d", 49805440634311326, 'http://example.com/lenna-512.png', [{'team_id': 1}], 1.0)]
-                r = redis.get_client()
+                r = redis_client.get_client()
                 r.delete(f"image_1c63abe0-aeb4-4bac-8925-948b69c32d0d")
                 r.lpush(f"image_1c63abe0-aeb4-4bac-8925-948b69c32d0d", json.dumps({"body": {"hash_value": 49805440634311326}}))
                 mock_response = Mock()
@@ -133,7 +133,7 @@ class TestSyncSimilarityBlueprint(BaseTestCase):
         with patch('requests.post') as mock_post_request:
             with patch('app.main.lib.image_similarity.execute_command') as mock_db_response:
                 mock_db_response.return_value = [(1, "1c63abe0-aeb4-4bac-8925-948b69c32d0d", 49805440634311326, 'http://example.com/lenna-512.png', [{'team_id': 1}], 1.0)]
-                r = redis.get_client()
+                r = redis_client.get_client()
                 r.delete(f"image_1c63abe0-aeb4-4bac-8925-948b69c32d0d")
                 r.lpush(f"image_1c63abe0-aeb4-4bac-8925-948b69c32d0d", json.dumps({"body": {"hash_value": 49805440634311326}}))
                 mock_response = Mock()
@@ -171,7 +171,7 @@ class TestSyncSimilarityBlueprint(BaseTestCase):
                 with patch('app.main.lib.shared_models.video_model.tmkpy.query') as mock_query:
                     mock_query.return_value = (1.0,)
                     mock_db_response.return_value = [(1, "Y2hlY2stcHJvamVjdF9tZWRpYS02Mzc2ODQtdmlkZW8", 'http://example.com/chair-19-sd-bar.mp4', "f4cf", "78f84604-f4cf-4044-a261-5fdf0ac44b63", [{'team_id': 1}], [-1363.0159912109375, 252.60726928710938, 652.66552734375, 48.47494888305664, -12.226404190063477, -62.87214279174805, -11.51701545715332, -13.31611442565918, -2.3773577213287354, -9.220880508422852, 30.38682746887207, -10.805936813354492, 17.883710861206055])]
-                    r = redis.get_client()
+                    r = redis_client.get_client()
                     r.delete(f"video_Y2hlY2stcHJvamVjdF9tZWRpYS02Mzc2ODQtdmlkZW8")
                     r.lpush(f"video_Y2hlY2stcHJvamVjdF9tZWRpYS02Mzc2ODQtdmlkZW8", json.dumps({"body": {"folder": "f4cf", "filepath": "78f84604-f4cf-4044-a261-5fdf0ac44b63", "hash_value": [-1363.0159912109375, 252.60726928710938, 652.66552734375, 48.47494888305664, -12.226404190063477, -62.87214279174805, -11.51701545715332, -13.31611442565918, -2.3773577213287354, -9.220880508422852, 30.38682746887207, -10.805936813354492, 17.883710861206055]}}))
                     r.delete(f"audio_Y2hlY2stcHJvamVjdF9tZWRpYS02Mzc2ODQtdmlkZW8")
@@ -212,7 +212,7 @@ class TestSyncSimilarityBlueprint(BaseTestCase):
                 with patch('app.main.lib.shared_models.video_model.tmkpy.query') as mock_query:
                     mock_query.return_value = (1.0,)
                     mock_db_response.return_value = [(1, "Y2hlY2stcHJvamVjdF9tZWRpYS02Mzc2ODQtdmlkZW8", 'http://example.com/chair-19-sd-bar.mp4', "f4cf", "78f84604-f4cf-4044-a261-5fdf0ac44b63", [{'team_id': 1}], [-1363.0159912109375, 252.60726928710938, 652.66552734375, 48.47494888305664, -12.226404190063477, -62.87214279174805, -11.51701545715332, -13.31611442565918, -2.3773577213287354, -9.220880508422852, 30.38682746887207, -10.805936813354492, 17.883710861206055])]
-                    r = redis.get_client()
+                    r = redis_client.get_client()
                     r.delete(f"video_Y2hlY2stcHJvamVjdF9tZWRpYS02Mzc2ODQtdmlkZW8")
                     r.lpush(f"video_Y2hlY2stcHJvamVjdF9tZWRpYS02Mzc2ODQtdmlkZW8", json.dumps({"body": {"folder": "f4cf", "filepath": "78f84604-f4cf-4044-a261-5fdf0ac44b63", "hash_value": [-1363.0159912109375, 252.60726928710938, 652.66552734375, 48.47494888305664, -12.226404190063477, -62.87214279174805, -11.51701545715332, -13.31611442565918, -2.3773577213287354, -9.220880508422852, 30.38682746887207, -10.805936813354492, 17.883710861206055]}}))
                     r.delete(f"audio_Y2hlY2stcHJvamVjdF9tZWRpYS02Mzc2ODQtdmlkZW8")
